not working mem_test version

+++ src/mem_test/main.c
@@ -1,8 +1,7 @@
 #include "xbr820.h"
 
-typedef	volatile uint32_t*	TEST_PTR;
+typedef	volatile uint8_t*	TEST_PTR;
 TEST_PTR ptr = (TEST_PTR)(XBR820_SRAM_BASE + XBR820_SRAM_SIZE/2);
-
 int main() {
 	for (int i = 0; i < 256; i++) {
 		ptr[i] = i;
